requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-chroma
chromadb
pydantic
python-dotenv
